Rename: user-storage 디렉터리 cookie/user-storage.ts로 변경
